pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-pixel"
version = "0.1.0"
description = "Train a real AI model that fits in a single pixel."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "David Velton"}
]
dependencies = [
    "numpy>=1.21",
    "Pillow>=9.0",
]

[project.optional-dependencies]
viz = ["matplotlib>=3.5"]
dev = ["pytest>=7.0", "matplotlib>=3.5"]

[project.scripts]
ai-pixel = "aipixel.cli:main"

[tool.setuptools.packages.find]
include = ["aipixel*"]
